library-project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
gunicorn
psycopg2-binary
whitenoise
